+++ modules/src/image/fourier/wiener.c
@@ -55,8 +55,8 @@ static void radialkernel(Fimage kernel, Fsignal rad)
   for (x=-nx/2;x<(nx+1)/2;x++)
     for (y=-ny/2;y<(ny+1)/2;y++) {
       adr = (y<0?ny+y:y)*nx+(x<0?nx+x:x);
-      dist = (int) floor(rad->scale * sqrt(cx * (double) (x * x)
-					   + cy * (double) (y * y)) + .5);
+      dist = floor(rad->scale * sqrt(cx * (double) (x * x)
+				     + cy * (double) (y * y)) + .5);
       if (dist>=rad->size) kernel->gray[adr] = 0.;
       else kernel->gray[adr] = rad->values[dist];
     }
